Create a new function to estimate the thermal conductivity of layers based on the work of Haigh; Donazzi; and an extra one for 'bulk' cases.
At the moment there is a unused variable in the parameter file:
'fixed at top', get rid off it.

IMPORTANT: the properties defined in the layer sections are for SOLIDS
not bulk properties. This is, for the soil grains instead of the
soil as a mixture (grains+water+air). The bulk properties are
estimated within the program.